% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millionaire.R
\name{days}
\alias{days}
\title{days: how much money after x days}
\usage{
days(d, p = 2, m = 7000)
}
\arguments{
\item{d}{days of trading}

\item{p}{daily percentage profit - default: 2}

\item{m}{start money}
}
\description{
days: how much money after x days
}
\examples{
days(100,2,5000)
}
\keyword{money}
