% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millionaire.R
\name{goal}
\alias{goal}
\title{Functions to calculate how soon you'll be a millionaire}
\usage{
goal(m, p = 2, g = 1e+06)
}
\arguments{
\item{m}{start money}

\item{p}{daily percentage profit - default: 2}

\item{g}{goal, default 1 million}

\item{d}{days of trading}
}
\description{
These functions will make you rich
goal: how many days to reach your goal
days: how much money after x days
}
\examples{
goal(5000,2,100000)
days(100,2,5000)
}
\keyword{money}
