% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedata.R
\name{mergedata}
\alias{mergedata}
\title{Function to merge different datasets}
\usage{
mergedata()
}
\description{
This function reads data, merges it into one dataset and creates output
}
\examples{
mergedata()
}
\keyword{data}
\keyword{merge}
