% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pidec.R
\name{pidec}
\alias{pidec}
\title{Function to calculate pi}
\usage{
pidec(i)
}
\arguments{
\item{i}{number of iterations in the G-L series}
}
\description{
Algorithm: Gregory-Leibniz
Note that results are incorrect beyond 15 decimals (R uses finite precision arithmetic)
}
\examples{
print(pidec(2000), digits = 15)
}
\keyword{gregory-leibniz}
\keyword{pi}
