% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propData.R
\name{propData}
\alias{propData}
\title{Function to calculate proportions in tables/dataframes}
\usage{
propData(data, byRow = TRUE, round = 2)
}
\arguments{
\item{data}{table/dataframe (no tibbles)}

\item{byRow}{calculate proportions by row? Default is TRUE}

\item{round}{number of decimal places}
}
\description{
This function uses frequencies as input and generates proportions (by row or column) as output
}
\examples{
propData(mytable, byRow=TRUE)
propData(mytable, byRow=FALSE)
}
\keyword{proportions}
