% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steyx.R
\name{steyx}
\alias{steyx}
\title{STEYX function}
\usage{
steyx(x, y)
}
\arguments{
\item{x}{known x}

\item{y}{known y}
}
\description{
Translation of STEYX function Excel > R
Returns the standard error of the predicted y-value for each x in the regression.
http://office.microsoft.com/en-au/excel-help/steyx-function-HP010062545.aspx
}
\examples{
steyx(10,150)
}
\keyword{steyx}
