% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vangle.R
\name{vangle}
\alias{vangle}
\title{Calculate viewing angle}
\usage{
vangle(a, b)
}
\arguments{
\item{a}{interstimulusdistance}

\item{b}{viewing distance}
}
\description{
This function calculates the viewing angle for 2-stimuli presentation experiments
}
\examples{
vangle(10,150)
}
\keyword{angle}
\keyword{viewing}
